requirements.txt
alembic==1.0.7
autopep8==1.4.3
Babel==2.6.0
blinker==1.4
Click==7.0
dominate==2.3.5
elasticsearch==6.3.1
Flask==1.0.2
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Moment==0.7.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.1
pep8==1.7.1
pkg-resources==0.0.0
pycodestyle==2.5.0
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2018.9
six==1.12.0
SQLAlchemy==1.2.18
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
